address review comments
- AddressManager: drop a stale X-address-derived destination tag when an invalid
  X-address is entered or a tagged X-address is replaced with an untagged one,
  so a tag can't ride onto a different destination (a hand-typed tag is kept).
- RippleHelper: serialize Sequence/LastLedgerSequence as Long — org.json can't
  wrap a Kotlin ULong value class, which would drop them from the rawJSON.
- Tests: requireNotNull instead of !! ; add tagged->untagged X-address regression.

+++ data/src/main/kotlin/com/vultisig/wallet/data/chains/helpers/RippleHelpter.kt
@@ -125,8 +125,10 @@ object RippleHelper {
                 "Destination" to keysignPayload.toAddress,
                 "Amount" to keysignPayload.toAmount.toString(),
                 "Fee" to gas.toString(),
-                "Sequence" to sequence,
-                "LastLedgerSequence" to lastLedgerSequence,
+                // org.json can't wrap a Kotlin ULong (a value class), so these would be dropped
+                // from the JSON — pass Long so Sequence/LastLedgerSequence serialize as numbers.
+                "Sequence" to sequence.toLong(),
+                "LastLedgerSequence" to lastLedgerSequence.toLong(),
                 "Memos" to
                     listOf(
                         mapOf(
